Restructured a couple things and cleaned up some code.
The game module is now in its own file. The file setup.js now just creates the global m3 file and its "sub-containers".

There is now a types directory, which should hold definitions for objects that we create instances of (as opposed to modules that are singular objects). Currently it only contains the Vector type, which was previously in the math module.