Fixed firefox bug with level backgrounds. One of the images wasn't getting loaded because it was corrupted or something.
In trying to fix the bug I created a new assets module which is just an object that holds all of our assets. The object gets created before the game starts, and the game waits to start until all of the assets have been loaded, so that the game doesn't start before all the images, etc. are loaded.

I hooked up all of our current assets to the module, and made a few other minor changes.